+++ test/solvers/test_primal_dual_semismooth_Newton.jl
@@ -18,9 +18,9 @@ using ManifoldDiff: differential_shortest_geodesic_startpoint, prox_distance
     δ = min(α / distance(pixelM, data[1], data[end]), 0.5)
     x_hat = shortest_geodesic(M, data, reverse(data), δ)
     N = M
-    fidelity(M, x) = 1 / 2 * distance(M, x, f)^2
-    Λ(M, x) = ArrayPartition(x, forward_logs(M, x))
-    prior(M, x) = norm(norm.(Ref(M.manifold), x, submanifold_component(N, Λ(x), 2)), 1)
+    fidelity(M, x) = 1 / 2 * distance(M, x, data)^2
+    Λ(M, x) = ArrayPartition(x, Manopt.Test.forward_logs(M, x))
+    prior(M, x) = norm(norm.(Ref(M.manifold), x, submanifold_component(N, Λ(M, x), 2)), 1)
     f(M, x) = (1 / α) * fidelity(M, x) + prior(M, x)
     prox_f(M, λ, x) = prox_distance(M, λ / α, data, x, 2)